refactor: Reorganize project-clone internal package to separate git functionality
